+++ ibis/sql/tests/test_compiler.py
@@ -1,3 +1,4 @@
+import datetime
 import unittest
 
 import pytest
@@ -2416,4 +2417,54 @@ WHERE t1.`a` = (
   SELECT max(`a`) AS `blah`
   FROM t1
 )"""
+    assert result == expected
+
+
+def test_table_drop_with_filter():
+    left = ibis.table(
+        [('a', 'int64'), ('b', 'string'), ('c', 'timestamp')],
+        name='t',
+    ).relabel({'c': 'C'})
+    left = left.filter(left.C == datetime.date(2018, 1, 1))
+    left = left.drop(['C'])
+    left = left.mutate(the_date=datetime.date(2018, 1, 1))
+
+    right = ibis.table([('b', 'string')], name='s')
+    joined = left.join(right, left.b == right.b)
+    joined = joined[left.a]
+    joined = joined.filter(joined.a < 1.0)
+    result = to_sql(joined)
+    # previously this was generating incorrect aliases due to not binding the
+    # self to expressions when calling projection:
+    # SELECT t0.`a`
+    # FROM (
+    #   SELECT `a`, `b`, '2018-01-01' AS `the_date`
+    #   FROM (
+    #     SELECT *
+    #     FROM (
+    #       SELECT `a`, `b`, `c` AS `C`
+    #       FROM t
+    #     ) t3
+    #     WHERE `C` = '2018-01-01'
+    #   ) t2
+    # ) t0
+    #   INNER JOIN s t1
+    #     ON t0.`b` = t1.`b`
+    # WHERE t0.`a` < 1.0
+    expected = """\
+SELECT t0.`a`
+FROM (
+  SELECT `a`, `b`, '2018-01-01' AS `the_date`
+  FROM (
+    SELECT *
+    FROM (
+      SELECT `a`, `b`, `c` AS `C`
+      FROM t
+    ) t3
+    WHERE `C` = '2018-01-01'
+  ) t2
+) t0
+  INNER JOIN s t1
+    ON t0.`b` = t1.`b`
+WHERE t0.`a` < 1.0"""
     assert result == expected
